+++ cmd/oras/internal/display/content/discard.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package raw
+package content
 
 import v1 "github.com/opencontainers/image-spec/specs-go/v1"
