+++ cmd/oras/internal/display/content/interface.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package raw
+package content
 
 import (
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
