+++ cmd/oras/internal/display/content/manifest_fetch.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package raw
+package content
 
 import (
 	"bytes"
